Move back ParseID to zedlake cmd
ParseID should be in the zedlake cmd package- it doesn't belong in package
lake. The service will expect that ksuid's are base 62 encoded.
