--- lake/id.go
@@ -1,31 +1,0 @@
-package lake
-
-import (
-	"encoding/hex"
-	"fmt"
-
-	"github.com/segmentio/ksuid"
-)
-
-func ParseID(s string) (ksuid.KSUID, error) {
-	// Check if this is a cut-and-paste from ZNG, which encodes
-	// the 20-byte KSUID as a 40 character hex string with 0x prefix.
-	var id ksuid.KSUID
-	if len(s) == 42 && s[0:2] == "0x" {
-		b, err := hex.DecodeString(s[2:])
-		if err != nil {
-			return ksuid.Nil, fmt.Errorf("illegal hex tag: %s", s)
-		}
-		id, err = ksuid.FromBytes(b)
-		if err != nil {
-			return ksuid.Nil, fmt.Errorf("illegal hex tag: %s", s)
-		}
-	} else {
-		var err error
-		id, err = ksuid.Parse(s)
-		if err != nil {
-			return ksuid.Nil, fmt.Errorf("%s: invalid commit ID", s)
-		}
-	}
-	return id, nil
-}
